fixed setUp()

+++ src/main/java/master/eit/manager/ZooManager.java
@@ -49,37 +49,39 @@ public class ZooManager implements Runnable {
 
     public void setUp() throws KeeperException, InterruptedException {
 
-        Stat enroll_exist = zoo.exists("/request/enroll", true);
-        if ( enroll_exist != null) {
-            deleteWithChildreen("/request/enroll", enroll_exist.getVersion());
-            zoo.create("/request/enroll", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
-        } else {
-            zoo.create("/request/enroll", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
+        Stat request_exist = zoo.exists("/request", true);
+        
+        if ( request_exist != null) {
+
+
+            Stat enroll_exist = zoo.exists("/request/enroll", true);
+            if ( enroll_exist != null) {
+                deleteWithChildreen("/request/enroll", enroll_exist.getVersion());
+            }
+
+            Stat quit_exist = zoo.exists("/request/quit", true);
+            if ( quit_exist != null) {
+                deleteWithChildreen("/request/quit", quit_exist.getVersion());
+            }
+            deleteWithChildreen("/request", request_exist.getVersion());
         }
 
-        Stat quit_exist = zoo.exists("/request/quit", true);
-        if ( quit_exist != null) {
-            deleteWithChildreen("/request/quit", quit_exist.getVersion());
-            zoo.create("/request/quit", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
-        } else {
-            zoo.create("/request/quit", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
-        }
-
+        
         Stat registry_exist = zoo.exists("/registry", true);
         if ( registry_exist != null) {
             deleteWithChildreen("/registry", registry_exist.getVersion());
-            zoo.create("/registry", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
-        } else {
-            zoo.create("/registry", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
         }
 
         Stat online_exist = zoo.exists("/online", true);
         if ( online_exist != null) {
             deleteWithChildreen("/online", online_exist.getVersion());
-            zoo.create("/online", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
-        } else {
-            zoo.create("/online", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
         }
+            
+        zoo.create("/request", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
+        zoo.create("/request/enroll", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
+        zoo.create("/request/quit", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
+        zoo.create("/registry", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
+        zoo.create("/online", null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
     }
 
     public void deleteWithChildreen (String path, int version) {
